+++ src/petaly/connectors/aws/redshift/rs_extractor.py
@@ -1,47 +1,78 @@
-import logging
+# Copyright © 2024-2025 Pavel Rabaev
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
+import logging
+logger = logging.getLogger(__name__)
+
+import sys
+import os
 from petaly.core.db_extractor import DBExtractor
 from petaly.utils.utils import FormatDict
 
-from petaly.connectors.aws.redshift.rs_connector import RSConnector
+from petaly.connectors.aws.redshift.rs_connector import RSConnectorIAM, RSConnectorTCP
+from petaly.connectors.aws.s3.s3_connector import S3Connector
 
 
 class RSExtractor(DBExtractor):
     def __init__(self, pipeline):
         #connection_params = self.get_connection_params(pipeline.source_attr)
-        self.db_connector = RSConnector(pipeline.source_attr)
+
+        if pipeline.source_attr.get('connection_method') == 'iam':
+            self.db_connector = RSConnectorIAM(pipeline.source_attr)
+            self.s3_connector = S3Connector(pipeline.source_attr, self.db_connector.aws_session)
+        elif pipeline.source_attr.get('connection_method') == 'tcp':
+            self.db_connector = RSConnectorTCP(pipeline.source_attr)
+            self.s3_connector = S3Connector(pipeline.source_attr, aws_session=None)
+        else:
+            logger.error(f"The connection_method: {pipeline.source_attr.get('connection_method')} is not supported for AWS extraction.")
+            sys.exit()
 
         super().__init__(pipeline)
         self.cloud_bucket_name = self.pipeline.source_attr.get('aws_bucket_name')
-        self.cloud_bucket_path = self.db_connector.bucket_prefix + self.cloud_bucket_name + '/'
-        self.cloud_iam_role = self.pipeline.source_attr.get('aws_iam_role')
+        self.cloud_bucket_path = self.s3_connector.bucket_prefix + self.cloud_bucket_name
+        self.aws_iam_role = self.pipeline.source_attr.get('aws_iam_role')
 
     def extract_data(self):
         super().extract_data()
 
     def get_query_result(self, meta_query):
-        return self.db_connector.get_query_result(meta_query)
+        return self.db_connector.get_metaquery_result(meta_query)
 
     def extract_to(self, extractor_obj_conf):
         object_name = extractor_obj_conf.get('object_name')
         extract_to_stmt = extractor_obj_conf.get('extract_to_stmt')
         logging.info(f"Statement to execute:{extract_to_stmt}")
 
+        blob_prefix = (self.pipeline.pipeline_name).strip('/')
         # cleanup object from s3 bucket
-        self.db_connector.drop_object_from_bucket(self.cloud_bucket_name, object_name)
+        self.s3_connector.drop_object_from_bucket(self.cloud_bucket_name, blob_prefix, object_name)
 
         # export data into s3 bucket
         self.db_connector.extract_to(extract_to_stmt)
 
         # download export from s3 into local folder
-        self.db_connector.download_files_from_bucket(self.cloud_bucket_name, object_name, self.pipeline.output_pipeline_dpath)
+        output_data_object_dir=extractor_obj_conf.get('output_data_object_dir')
+
+        self.s3_connector.download_object_from_bucket(self.cloud_bucket_name, blob_prefix, object_name, output_data_object_dir)
 
     def compose_extract_to_stmt(self, extract_to_stmt, extractor_obj_conf) -> dict:
         """ Its save copy statement into file
         """
         extract_data_options = self.compose_extract_options(extractor_obj_conf)
         object_name = extractor_obj_conf.get('object_name')
-        extract_to_fpath = self.cloud_bucket_path + object_name  + '/' + self.pipeline.data_dname + '/' + object_name + '_'
+
+        extract_to_fpath = (self.cloud_bucket_path + '/' + self.pipeline.pipeline_name + '/' + object_name  + '/' + object_name + '_').strip('/')
 
         extract_to_stmt = extract_to_stmt.format_map(
         					FormatDict( column_list=extractor_obj_conf.get('column_list'),
@@ -49,7 +80,7 @@ class RSExtractor(DBExtractor):
                                         table_name=extractor_obj_conf.get('source_object_name'),
                                         extract_to_fpath = extract_to_fpath,
                                         extract_to_options=extract_data_options,
-                                        iam_role=self.cloud_iam_role
+                                        iam_role=self.aws_iam_role
                                        ))
 
         return extract_to_stmt