+++ src/petaly/connectors/aws/redshift/rs_loader.py
@@ -1,19 +1,50 @@
+# Copyright © 2024-2025 Pavel Rabaev
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import logging
+logger = logging.getLogger(__name__)
+
+import sys
+
 from petaly.utils.utils import FormatDict
 from petaly.core.db_loader import DBLoader
 
-from petaly.connectors.aws.redshift.rs_connector import RSConnector
+#from petaly.connectors.aws.redshift.rs_connector import RSConnector
+from petaly.connectors.aws.redshift.rs_connector import RSConnectorIAM, RSConnectorTCP
+from petaly.connectors.aws.s3.s3_connector import S3Connector
 
 
 class RSLoader(DBLoader):
 
     def __init__(self, pipeline):
         #connection_params = self.get_connection_params(pipeline.target_attr)
-        self.db_connector = RSConnector(pipeline.target_attr)
+
+        if pipeline.target_attr.get('connection_method') == 'iam':
+            self.db_connector = RSConnectorIAM(pipeline.target_attr)
+            self.s3_connector = S3Connector(pipeline.source_attr, self.db_connector.aws_session)
+        elif pipeline.target_attr.get('connection_method') == 'tcp':
+            self.db_connector = RSConnectorTCP(pipeline.target_attr)
+            self.s3_connector = S3Connector(pipeline.source_attr, aws_session=None)
+        else:
+            logger.error(f"The connection_method: {pipeline.source_attr.get('connection_method')} is not supported for AWS load.")
+            sys.exit()
+
 
         super().__init__(pipeline)
         self.cloud_bucket_name = self.pipeline.target_attr.get('aws_bucket_name')
-        self.cloud_bucket_path = self.db_connector.bucket_prefix + self.cloud_bucket_name + '/'
-        self.cloud_iam_role = self.pipeline.target_attr.get('aws_iam_role')
+        self.cloud_bucket_path = self.s3_connector.bucket_prefix + self.cloud_bucket_name
+        self.aws_iam_role = self.pipeline.target_attr.get('aws_iam_role')
 
     def load_data(self):
         super().load_data()
@@ -35,8 +66,9 @@ class RSLoader(DBLoader):
     def load_from(self, loader_obj_conf):
 
         object_name = loader_obj_conf.get('object_name')
+        blob_prefix = (self.pipeline.pipeline_name + '/' + object_name).strip('/')
         # 1. cleanup object from bucket
-        self.db_connector.drop_object_from_bucket(self.cloud_bucket_name, object_name)
+        self.s3_connector.drop_object_from_bucket(self.cloud_bucket_name, blob_prefix, object_name)
 
         # 2. drop and recreate table
         if loader_obj_conf.get('recreate_destination_object') == True:
@@ -49,12 +81,12 @@ class RSLoader(DBLoader):
 
         file_list = self.f_handler.get_specific_files(output_data_object_dir, '*.csv.gz')
 
-        self.db_connector.load_files_to_s3_bucket(self.cloud_bucket_name, object_name, file_list)
-        s3_file_list = self.db_connector.get_bucket_file_list(self.cloud_bucket_name, object_name)
+        self.s3_connector.load_files_to_s3_bucket(self.cloud_bucket_name, blob_prefix, object_name, file_list)
+        s3_file_list = self.s3_connector.get_bucket_file_list(self.cloud_bucket_name, blob_prefix, object_name)
         load_from_stmt = loader_obj_conf.get('load_from_stmt')
 
         for data_fpath in s3_file_list:
-            path_to_data_file = self.cloud_bucket_path + data_fpath
+            path_to_data_file = self.cloud_bucket_path + '/' + self.pipeline.pipeline_name + '/' + data_fpath
             load_from_stmt = load_from_stmt.format_map(
                 FormatDict(path_to_data_file=path_to_data_file))
 
@@ -112,12 +144,12 @@ class RSLoader(DBLoader):
 
         load_from_stmt = load_from_stmt.format_map(FormatDict(schema_table_name=schema_table_name,
                                                                column_list=column_list,
-                                                              iam_role=self.cloud_iam_role,
+                                                              iam_role=self.aws_iam_role,
                                                               load_from_options=load_data_options))
         load_from_file_fpath = loader_obj_conf.get('load_from_stmt_fpath')
         self.f_handler.save_file(load_from_file_fpath, load_from_stmt)
 
         return load_from_stmt
 
-    def cleanup_object_relations(self, object_name):
-        self.db_connector.drop_object_from_bucket(self.cloud_bucket_name, object_name)
+    def cleanup_object_relations(self, blob_prefix, object_name):
+        self.s3_connector.drop_object_from_bucket(self.cloud_bucket_name, blob_prefix, object_name)