+++ src/petaly/connectors/aws/s3/s3_extractor.py
@@ -1,4 +1,4 @@
-# Copyright © 2024 Pavel Rabaev
+# Copyright © 2024-2025 Pavel Rabaev
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -13,24 +13,23 @@
 # limitations under the License.
 
 import logging
+import sys
+
 logger = logging.getLogger(__name__)
 
 import os
 import logging
 
-from petaly.connectors.gcp.gs.gs_connector import GSConnector
+from petaly.connectors.aws.s3.s3_connector import S3Connector
 from petaly.core.f_extractor import FExtractor
 from petaly.utils.file_handler import FileHandler
-from petaly.utils.utils import FormatDict
 
-class GSExtractor(FExtractor):
+
+class S3Extractor(FExtractor):
     def __init__(self, pipeline):
-        self.gs_connector = GSConnector()
+        self.s3_connector = S3Connector(pipeline.source_attr, aws_session=None)
         super().__init__(pipeline)
-        self.cloud_bucket_name = self.pipeline.source_attr.get('gcp_bucket_name')
-        self.cloud_project_id = self.pipeline.source_attr.get('gcp_project_id')
-        self.cloud_region = self.pipeline.source_attr.get('gcp_region')
-
+        self.cloud_bucket_name = self.pipeline.source_attr.get('aws_bucket_name')
         self.f_handler = FileHandler()
 
     def extract_data(self):
@@ -43,6 +42,7 @@ class GSExtractor(FExtractor):
 
         object_source_dir = extractor_obj_conf.get('object_source_dir')
         blob_prefix = object_source_dir.strip('/')
+
         file_names = extractor_obj_conf.get('file_names')
 
         file_names_full_path = []
@@ -50,12 +50,12 @@ class GSExtractor(FExtractor):
             file_names_full_path.append(f"{blob_prefix}/{file_name}")
 
         # download export from bucket into local folder
-        downloaded_file_list = self.gs_connector.download_files_from_bucket(
+        downloaded_file_list = self.s3_connector.download_files_from_bucket(
                                                     self.cloud_bucket_name,
-                                                    blob_prefix,
-                                                    specific_file_list=file_names_full_path,
+                                                    file_names=file_names,
                                                     destination_directory=output_data_object_dir)
 
         logging.debug(f"Following file list were downloaded from bucket:\n{downloaded_file_list}")
+