+++ src/petaly/connectors/postgres/config/extract_to_stmt.sql
@@ -1,3 +1,6 @@
 COPY (SELECT {column_list}
 FROM {schema_name}.{table_name}) TO STDOUT
-WITH (FORMAT CSV {copy_to_options});
+WITH (
+FORMAT CSV
+{copy_to_options}
+);