+++ src/petaly/connectors/postgres/config/load_from_stmt.sql
@@ -1,5 +1,8 @@
 COPY {schema_table_name}
 {column_list}
 FROM STDIN
-WITH (FORMAT CSV {copy_from_options})
+WITH (
+FORMAT CSV
+{copy_from_options}
+)
 ;