+++ src/petaly/core/db_extractor.py
@@ -109,12 +109,12 @@ class DBExtractor(ABC):
 
 		extractor_obj_conf = {'object_name': object_name}
 
-		# 1. compose output data dir
+		# 1. compose metadata directory
 		output_metadata_object_dir = self.pipeline.output_object_metadata_dpath.format(object_name=object_name)
 		extractor_obj_conf.update({'output_metadata_object_dir': output_metadata_object_dir})
+		metadata_fpath = self.pipeline.output_object_metadata_fpath.format(object_name=object_name)
 
 		# 2. get and compose metadata query
-		metadata_fpath = self.pipeline.output_object_metadata_fpath.format(object_name=object_name)
 		table_metadata = self.f_handler.load_file_as_dict(metadata_fpath, 'json')
 		extract_queries_dict = self.compose_extract_queries(table_metadata)
 		extractor_obj_conf.update(extract_queries_dict)
@@ -135,14 +135,19 @@ class DBExtractor(ABC):
 		extractor_obj_conf.update({'extract_to_stmt_fpath': output_extract_to_stmt_fpath})
 		self.f_handler.save_file(output_extract_to_stmt_fpath, extract_to_stmt)
 
-		# 6. create output_file_path
-		output_object_fpath = self.get_local_output_path(object_name)
+		# 6. compose output data dir
+		output_data_object_dir = self.pipeline.output_object_data_dpath.format(object_name=object_name)
+		extractor_obj_conf.update({'output_data_object_dir': output_data_object_dir})
+		self.f_handler.make_dirs(output_data_object_dir)
+
+		# compose output object file path
+		output_object_fpath = os.path.join(output_data_object_dir, object_name + '.csv')
 		extractor_obj_conf.update({'output_object_fpath': output_object_fpath})
 
 		logger.debug(f"Config for data extract: {extractor_obj_conf}")
 		return extractor_obj_conf
 
-	def get_local_output_path(self, object_name):
+	def deprecated_get_local_output_path(self, object_name):
 
 		object_dpath = self.pipeline.output_object_data_dpath.format(object_name=object_name)